Stop the login interface from sticking to the back after switching windows
Moved both class [MainWindow] and [Login] to the same file to avoid circular imports when I implement a log out function.
The closeWindow functions destroys the frame1 but doesn't exactly solve the issue, it won't work for main window as it is right now. A work-around I thought of is adding a main frame that holds
all other frames and destroy it when logging out, but I think that will force both windows to be of the same size.